refactor licensing logic to gumroad
