+++ airsync-mac/Core/AppState.swift
@@ -58,7 +58,7 @@ class AppState: ObservableObject {
         }
 
         Task {
-            await checkLicenseIfNeeded()
+            await Gumroad().checkLicenseIfNeeded()
         }
 
         self.scrcpyBitrate = UserDefaults.standard.integer(forKey: "scrcpyBitrate")
@@ -564,155 +564,6 @@ class AppState: ObservableObject {
         }
     }
 
-    func checkLicenseIfNeeded() async {
-        // If we already had a successful check today, skip to enforce "max one successful check per day"
-        if let lastSuccess = UserDefaults.standard.lastLicenseSuccessfulCheckDate,
-           Calendar.current.isDateInToday(lastSuccess) {
-            print("License already successfully validated today — skipping network call.")
-            return
-        }
-
-        await checkLicense()
-    }
-
-    @MainActor
-    func checkLicense() async {
-        // Always record that we attempted today (used to prevent double-counting network failures in one day)
-        let now = Date()
-        let calendar = Calendar.current
-
-        // If no stored key, behave as before
-        guard let key = self.licenseDetails?.key, !key.isEmpty else {
-            self.isPlus = false
-            incrementInvalidLicenseFailCount() // treat as invalid (no key)
-            UserDefaults.standard.lastLicenseCheckDate = now
-            return
-        }
-
-        do {
-            let valid = try await checkLicenseKeyValidity(
-                key: key,
-                save: false,
-                isNewRegistration: false
-            )
-
-            UserDefaults.standard.lastLicenseCheckDate = now
-
-            if valid {
-                // Successful validation today
-                UserDefaults.standard.consecutiveNetworkFailureDays = 0
-                UserDefaults.standard.consecutiveLicenseFailCount = 0
-                UserDefaults.standard.lastLicenseSuccessfulCheckDate = now
-                self.isPlus = true
-                print("License valid — daily success recorded.")
-            } else {
-                // Invalid/expired/cancelled/license-limit — disable immediately
-                self.isPlus = false
-                incrementInvalidLicenseFailCount()
-                // Reset network failure streak because this is not a network failure
-                UserDefaults.standard.consecutiveNetworkFailureDays = 0
-                print("License invalid or expired — disabled Plus.")
-            }
-        } catch let error as LicenseCheckError {
-            // Network/server failure: do not disable Plus today
-            UserDefaults.standard.lastLicenseCheckDate = now
-
-            // Only increment once per calendar day
-            var consecutiveDays = UserDefaults.standard.consecutiveNetworkFailureDays
-            if UserDefaults.standard.lastLicenseSuccessfulCheckDate != nil {
-                // last successful check date exists; if it's today we already validated; but we are here only if not successful today
-                // we still want to count by day against previous attempt date
-            }
-            // Compare with the last attempt day (not last success) to avoid double-counting same day
-            if UserDefaults.standard.lastLicenseCheckDate != nil {
-                // We just set it to 'now'; we need the previous value to compare. To avoid this race,
-                // check by reading previous date first before setting in future refactors.
-                // For current code path, guard by storing previous date before the call if needed.
-            }
-            // Simpler: increment if last success date is not today OR there was no success date recently
-            // Also ensure we don't increment more than once a day by comparing with a stored "last network failure day" if needed.
-            // For simplicity, we’ll increment if not already incremented today:
-            let lastNetworkDay = UserDefaults.standard.object(forKey: "lastNetworkFailureDay") as? Date
-            if lastNetworkDay == nil || !calendar.isDate(lastNetworkDay!, inSameDayAs: now) {
-                consecutiveDays += 1
-                UserDefaults.standard.set(now, forKey: "lastNetworkFailureDay")
-            }
-            UserDefaults.standard.consecutiveNetworkFailureDays = consecutiveDays
-
-            // User messaging
-            postNativeNotification(
-                id: "license_network_issue",
-                appName: "AirSync+",
-                title: "License check skipped",
-                body: "Network issue while validating your license. \(consecutiveDays)/3 consecutive days."
-            )
-
-            if consecutiveDays >= 3 {
-                // Unregister on 3rd consecutive day
-                performUnregisterWithAlert(reason: "Could not validate your license for 3 consecutive days due to network issues. Please re-enter your key when you’re online.")
-            } else {
-                print("Network/server error during license check: \(error)")
-            }
-        } catch {
-            // Any other unexpected error — treat as server error category
-            UserDefaults.standard.lastLicenseCheckDate = now
-
-            let lastNetworkDay = UserDefaults.standard.object(forKey: "lastNetworkFailureDay") as? Date
-            if lastNetworkDay == nil || !Calendar.current.isDate(lastNetworkDay!, inSameDayAs: now) {
-                let newVal = UserDefaults.standard.consecutiveNetworkFailureDays + 1
-                UserDefaults.standard.consecutiveNetworkFailureDays = newVal
-                UserDefaults.standard.set(now, forKey: "lastNetworkFailureDay")
-            }
-
-            postNativeNotification(
-                id: "license_network_issue",
-                appName: "AirSync+",
-                title: "License check skipped",
-                body: "A server error occurred while validating your license."
-            )
-
-            if UserDefaults.standard.consecutiveNetworkFailureDays >= 3 {
-                performUnregisterWithAlert(reason: "Could not validate your license for 3 consecutive days due to server issues. Please re-enter your key when you’re online.")
-            } else {
-                print("Unexpected error during license check: \(error)")
-            }
-        }
-    }
-
-    private func incrementInvalidLicenseFailCount() {
-        let failCount = UserDefaults.standard.consecutiveLicenseFailCount + 1
-        UserDefaults.standard.consecutiveLicenseFailCount = failCount
-
-        if failCount >= 3 {
-            clearLicenseDetails()
-            print("License check failed \(failCount) times — license removed")
-        }
-    }
-
-    private func performUnregisterWithAlert(reason: String) {
-        // Clear local license and disable Plus
-        self.isPlus = false
-        clearLicenseDetails()
-        UserDefaults.standard.consecutiveNetworkFailureDays = 0
-        UserDefaults.standard.set(nil, forKey: "lastNetworkFailureDay")
-
-        // Inform user with a blocking popup
-        DispatchQueue.main.async {
-            let alert = NSAlert()
-            alert.alertStyle = .warning
-            alert.addButton(withTitle: "OK")
-            alert.messageText = "AirSync+ Unregistered"
-            alert.informativeText = reason
-            alert.runModal()
-        }
-    }
-
-    private func clearLicenseDetails() {
-        self.licenseDetails = nil
-        UserDefaults.standard.removeObject(forKey: "licenseDetailsKey")
-        UserDefaults.standard.consecutiveLicenseFailCount = 0
-    }
-
     func updateDockIconVisibility() {
         DispatchQueue.main.async {
             if self.hideDockIcon {
@@ -723,4 +574,3 @@ class AppState: ObservableObject {
         }
     }
 }
-
